pyproject.toml
[tool.poetry]
name = "job-parser-db-mvp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "job_parser_db_mvp"}]

[tool.poetry.dependencies]
python = "^3.11"
certifi = "^2023.7.22"
charset-normalizer = "^3.2.0"
idna = "^3.4"
psycopg2 = "^2.9.7"
requests = "^2.31.0"
urllib3 = "^2.0.5"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
